pm: handle reset prep IRQ
In the SoC warm reboot process, sometimes there are
instances where the ISH firmware fails to receive the
warm reboot request from PMC, leading to the failure to reboot
ISH firmware hence causing ISHTP kernel driver communication
failure with ISH firmware. This issue is observred when
there is an IPC request pending at the ISH and at the time
a warm reboot is initiated causing IPM driver timeout.

The IPM driver timeout is set to 1sec waiting for the Peer ACK.
Due to warm reboot, Peer ACK is not received leading to timeout
of 1sec. During this time ISH HW triggeres the RESET
PREP IRQ, but we do not observe ISR routine exexuted at firmware
side which is expected to handle the warm reboot.

Changes done to perform ish pm init using zephyr SYS_INIT() in
PRE_KERNEL_2 stage. This change helps handle the IRQ rightly and
invokes the ISR routine to handle the firmware reset.

Change-Id: I454bd75257cdfc1c8f37e582231f6fe7f5e8f91d